% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callimbalance.R
\name{compute_imbalance_statistics}
\alias{compute_imbalance_statistics}
\title{Compute Imbalance Statistics}
\usage{
compute_imbalance_statistics(counts_frame, null_shape1, null_shape2,
  lsse_shape1, lsse_shape2, minimum_coverage = 10)
}
\arguments{
\item{counts_frame}{data frame containing allele counts}

\item{null_shape1}{first shape parameter of the null distribution
(from npbin)}

\item{null_shape2}{second shape parameter of the null distribution
(from npbin)}

\item{lsse_shape1}{first shape parameter of the LSSE estimate}

\item{lsse_shape2}{second shape parameter of the LSSE estimate}

\item{minimum_coverage}{minimum coverage level for filtering variants}
}
\value{
data frame, including allele counts and imbalance statistics
}
\description{
compute per-variant imbalance statistics, including p-values
  and effect sizes
}
