% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callimbalance.R
\name{compute_total_coverage}
\alias{compute_total_coverage}
\title{compute total coverage}
\usage{
compute_total_coverage(intermediate_frames)
}
\arguments{
\item{intermediate_frames}{list of intermediate data frames}
}
\value{
integer, the total coverage level
}
\description{
compute total coverage level across samples
}
