% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callimbalance.R
\name{compute_total_ref_count}
\alias{compute_total_ref_count}
\title{compute total ref count}
\usage{
compute_total_ref_count(intermediate_frames)
}
\arguments{
\item{intermediate_frames}{list of intermediate data frames}
}
\value{
integer, the total reference allel count
}
\description{
compute total reference allele count across samples
}
