% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callimbalance.R
\name{coverage_frame}
\alias{coverage_frame}
\title{coverage frame}
\usage{
coverage_frame(intermediate_frames)
}
\arguments{
\item{intermediate_frames}{list of intermediate data frames}
}
\value{
data frame, the coverage levels
}
\description{
data frame of coverage levels
}
