% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callimbalance.R
\name{meta_analyze}
\alias{meta_analyze}
\title{meta analyze imbalance statistics}
\usage{
meta_analyze(row)
}
\arguments{
\item{row}{a row of imbalance data}
}
\value{
meta-analyzed effect size
}
\description{
meta analyze imbalance statistics
}
