% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callimbalance.R
\name{ref_frame}
\alias{ref_frame}
\title{ref frame}
\usage{
ref_frame(intermediate_frames)
}
\arguments{
\item{intermediate_frames}{list of intermediate data frames}
}
\value{
data frame, the ref counts
}
\description{
data frame of reference allele counts
}
